added a config option for multi tenent edit_location, add_location, delete_location now work correctly edit_location, add_location, delete_location now return to the org view or list_loctions depending on multi tenent config option delete_location on has a pop up confirm box added get_org_id funciton to location model

+++ code_igniter/application/views/theme-tango/v_location.php
@@ -65,7 +65,7 @@
 				<a href="../../admin_location/edit_location/<?php echo $location->id?>"><img style='float: right; margin; 10px; ' src='<?php echo $oa_theme_images;?>/48_edit.png' alt='Edit Location' title='Edit Location' width='48'/></a>
 			</td></tr>
 			<tr><td>
-                		<a href="../../admin_location/delete_location/<?php echo $location->id?>"><img style='float: right; margin; 10px; ' src='<?php echo $oa_theme_images;?>/48_delete_page.png' alt='Delete Location' title='Delete Location' width='48'/></a>
+                		<a href="../../admin_location/delete_location/<?php echo $location->id?>" onclick="return confirm('Ok to Delete?')"><img style='float: right; margin; 10px; ' src='<?php echo $oa_theme_images;?>/48_delete_page.png' alt='Delete Location' title='Delete Location' width='48'/></a>
 			</td></tr>
 			<tr><td>
 		                <a href="../../admin_network/add_network/<?php echo $location->id?>"><img style='float: right; margin; 10px; ' src='<?php echo $oa_theme_images;?>/48_add_page.png' alt='Add Network' title='Add Network' width='48'/></a>
